+++ src/dal/SongsDAO.java
@@ -42,7 +42,7 @@ public class SongsDAO {
                 while (resultSet.next()) {
                     String title = resultSet.getString("title");
                     String artist = resultSet.getString("artist");
-                    double songLength = resultSet.getDouble("songLength");
+                    float songLength = resultSet.getFloat("songLength");
                     String category = resultSet.getString("category");
                     String url = resultSet.getString("url");
                     int ID = resultSet.getInt("ID");
@@ -59,13 +59,13 @@ public class SongsDAO {
     }
 
 
-    public Song createSong(String title, String artist, double songLength, String category, String url) throws SQLException {
+    public Song createSong(String title, String artist, float songLength, String category, String url) throws SQLException {
         String sql = "INSERT INTO SONG(Title, Artist, songLength, category, Url) values (?,?,?,?,?);";
         Connection connection = databaseConnector.getConnection();
         try (PreparedStatement preparedStatement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
             preparedStatement.setString(1, title);
             preparedStatement.setString(2, artist);
-            preparedStatement.setDouble(3, songLength);
+            preparedStatement.setFloat(3, songLength);
             preparedStatement.setString(4, category);
             preparedStatement.setString(5, url);
             preparedStatement.executeUpdate();
